replaced magic value in game listener with enum

+++ RockPaperScissors/Game.cs
@@ -32,7 +32,7 @@ namespace RockPaperScissors
 
             if (whoWins != PlayResult.Draw)
             {
-                listener.GameOver((int) whoWins);
+                listener.GameOver(whoWins);
             }
         }
